refactor: move createTeamsHandler business logic from trpc to features
